Changed how audit_world works with /nextRegion and /nextDeleteRegion
Used to be an issue with the server storing chunks in memory, even with /chunk unload, causing some chunks to not be copied over correctly, which could lead to it seeming like there is no progress when in reality there is.
This update makes it so the world is unloaded and re-generated.
This does as an effect mean that only 1 user is able to audit at a time